Search Container refactor
