+++ src/js/components/search/table/ResultsColumnOption.jsx
@@ -25,17 +25,19 @@ export default class ResultsColumnOption extends React.Component {
 
     render() {
         const checked = this.props.checked;
-        return (<li>
-            <input
-                type="checkbox"
-                id={`column-${this.props.column}`}
-                value={this.props.label}
-                checked={checked}
-                onChange={this.toggleColumnVisibility} />
-            <label htmlFor={`column-${this.props.column}`}>
-                {this.props.label}
-            </label>
-        </li>);
+        return (
+            <li>
+                <input
+                    type="checkbox"
+                    id={`column-${this.props.column}`}
+                    value={this.props.label}
+                    checked={checked}
+                    onChange={this.toggleColumnVisibility} />
+                <label htmlFor={`column-${this.props.column}`}>
+                    {this.props.label}
+                </label>
+            </li>
+        );
     }
 }
