+++ src/js/components/search/table/ResultsTable.jsx
@@ -6,8 +6,6 @@
 import React from 'react';
 import Immutable from 'immutable';
 
-import _ from 'lodash';
-
 import IBTable from 'components/sharedComponents/IBTable/IBTable';
 
 import ResultsTableGenericCell from './cells/ResultsTableGenericCell';
@@ -40,13 +38,14 @@ export default class ResultsTable extends React.PureComponent {
 
         this.rowClassName = this.rowClassName.bind(this);
         this.tableScrolled = this.tableScrolled.bind(this);
-        this.columnsChanged = this.columnsChanged.bind(this);
     }
 
     componentWillReceiveProps(nextProps) {
+        const currentType = nextProps.currentType;
+        const visibileCols = parseFloat(nextProps.columns.length);
         // update the data hash
         this.setState({
-            dataHash: nextProps.currentType
+            dataHash: `${currentType}-${visibileCols}`
         });
     }
 
@@ -61,23 +60,10 @@ export default class ResultsTable extends React.PureComponent {
             return true;
         }
         else if (nextProps.columns !== this.props.columns) {
+            // re-render if column visibility changed
             return true;
         }
         return false;
-    }
-
-    componentDidUpdate(prevProps) {
-        if (!_.isEqual(prevProps.columns, this.props.columns)) {
-            this.columnsChanged();
-        }
-    }
-
-    columnsChanged() {
-        this.setState({
-            dataHash: parseFloat(this.props.columns.length)
-        }, () => {
-            this.prepareTable();
-        });
     }
 
     tableScrolled(xPos, yPos) {
