+++ src/js/redux/reducers/index.js
@@ -25,8 +25,8 @@ const appReducer = combineReducers({
     resultsBatch: resultsBatchReducer,
     filters: filtersReducer,
     searchOrder: orderReducer,
-    columnVisibility: columnVisibilityReducer,
     autocompleteLocations: autocompleteReducer,
+    columnVisibility: columnVisibilityReducer,
     autocompleteAwardIDs: awardIDReducer,
     autocompleteAgencies: agencyReducer,
     autocompleteRecipients: recipientReducer,
